\name{workout.tracker-package}
\alias{workout.tracker-package}
\alias{workout.tracker}
\docType{package}
\title{
\packageTitle{workout.tracker}
}
\description{
\packageDescription{workout.tracker}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{workout.tracker}
\packageIndices{workout.tracker}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{workout.tracker}

Maintainer: \packageMaintainer{workout.tracker}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
